kospo_backend_python/pyproject.toml
[tool.poetry]
name = "kospo-main-service"
version = "0.1.0"
description = ""
authors = ["ccw7463 <ccw7463@example.com>"]
readme = "README.md"

packages = [
    { include = "scripts" },
    { include = "src" },
    { include = "tests" }
]

[tool.poetry.dependencies]
python = "^3.11"
joblib = "^1.3.2"
langchain = "^0.3.23"
langchain_community = "^0.3.21"
langchain_core = "^0.3.51"
langchain_huggingface = "^0.1.2"
langgraph = "^0.3.27"
matplotlib = "^3.8.2"
numpy = "^1.26.4"
pandas = "^1.5.3"
python-dotenv = "^1.1.0"
schedule = "^1.2.2"
scikit_learn = "^1.2.0"
seaborn = "^0.13.2"
shap = "^0.47.0"
sqlalchemy = "^1.4.54"
tqdm = "^4.66.1"
typing_extensions = "^4.13.1"
psycopg2-binary = "^2.9.10"
openai = "^1.98.0"
rich = "^14.1.0"
reportlab = "^4.4.3"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
pydantic = "^2.10.0"
pydantic-settings = "^2.2.1"
sse-starlette = "^1.8.2"
langchain-openai = "^0.3.28"
aiohttp = "^3.12.15"
redis = "^6.4.0"
faiss-cpu = "^1.11.0.post1"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[tool.poetry.scripts]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"